Pulled the identical tests out into a file to require by each database specific case.
