+++ spec/cases/each_with_ar_sqlite.rb
@@ -2,42 +2,11 @@ require './spec/cases/helper'
 require "active_record"
 require "sqlite3"
 STDOUT.sync = true
-in_worker_type = "in_#{ENV.fetch('WORKER_TYPE')}".to_sym
 
 ActiveRecord::Schema.verbose = false
 ENV["DATABASE_URL"] = "sqlite3:parallel_with_ar_test.sqlite3"
-ActiveRecord::Base.establish_connection
 
-class User < ActiveRecord::Base
-end
-
-# create tables
-unless User.table_exists?
-  ActiveRecord::Schema.define(:version => 1) do
-    create_table :users do |t|
-      t.string :name
-    end
-  end
-end
-
-User.delete_all
-
-3.times { User.create!(:name => "X") }
-
-print "Parent: "
-puts User.first.name
-
-
-# Run with both disabled and enabled Parallel (AR workarounds shouldn't break 0)
-[0,1].each do |zero_one|
-  print "Parallel (#{in_worker_type} => #{zero_one}): "
-  Parallel.each([1], in_worker_type => 0) do
-    puts User.all.map(&:name).join
-  end
-end
-
-print "\nParent: "
-puts User.first.name
+require "./spec/cases/each_with_ar_generic"
 
 # Delete the sqlite3 file.  :memory: was neat, but once you disconnect, it's just gone.
 `rm parallel_with_ar_test.sqlite3`
